Extract protein model design from IsoAPQModelDesign object

Extract protein model design from IsoAPQModelDesign object

Extract fixed effects model design from IsoAPQModelDesign object

Extract random model design from IsoAPQModelDesign object

Extract response vector from IsoAPQModelDesign object

Extract counts for levels of random effects from IsoAPQModelDesign object

Extract vector that describes independent blocks from IsoAPQModelDesign object

Check if model described by IsoAPQModelDesign has fixed effects

Check if model described by IsoAPQModelDesign has random effects

Extract model formula from IsoAPQModelDesign object

Extract number of random effects from IsoAPQModelDesign object

Extract number of fixed effects from IsoAPQModelDesign object

Extract number of proteins from IsoAPQModelDesign object

Extract number of observations from IsoAPQModelDesign object

Check if model described by IsoAPQModelDesign is linear

Get analytical gradient for a IsoAPQModelDesign object

Get negative log-likelihood for a IsoAPQModelDesign object

Fit a nonlinear mixed effects model based on a IsoAPQModelDesign object

getIsoProteinDesign(x)

getIsoFixedDesign(x)

getIsoRandomDesign(x)

getIsoResponse(x)

getIsoEffectsCounts(x)

getIsoIndependentUnit(x)

hasIsoFixedEffects(x)

hasIsoRandomEffects(x)

getIsoFormula(x)

getIsoNumRandom(x)

getIsoNumFixed(x)

getIsoNumProteins(x)

getIsoNumObservations(x)

isIsoLinear(x)

getIsoAnalyticalGradient(x)

getIsoLogLikelihood(x)

fitIsoModel(x, ...)

Arguments

x
object of class IsoAPQModelDesign

Value

matrix

matrix

matrix

numeric vector

numeric vector

character vecotr

logical

logical

formula

integer

integer

integer

integer

logical

function of a single parameter

function of a single parameter

object of class IsoAPQModel