S4 class to represent a nonlinear mixed effects model based on isotopic distribution

S4 class to represent a nonlinear mixed effects model based on isotopic distribution

Fit a nonlinear mixed effects model based on isotopic distribution

IsoAPQModel(model_formula, model_data, ...)

Arguments

model_formula
formula that specifies the model. Special notation `..` denotes all protein variables

model_data
input data - data.table with columns: intensity, iso_prob or isoDistr, sequence, rep, precursor_scan, charge_in_scan and protein indicator columns.

additional
parameters to the model fitting function (currently stats::optim)

Value

object of class IsoAPQModel

Slots

model_design

object of class IsoAPQModelDesign that describes experimental design reflected in the model

fitted_model_raw

usually list (output of stats::optim), potentially could be a different object with raw outputs of function used to fit a model

coef

numeric vector of model coefficient in the following order: standard deviation of the error, parameters that describe random effects, protein abundances, fixed effects

sigma

standard deviation of the random error

random_effects_var

numeric vector of estimated variances of random effects

protein_abundance

numeric vector of estimated protein abundances

fixed_effects

numeric vector of estimated fixed effects

fitted

numeric vector of fitted values

residuals

numeric vector of residuals

hessian

hessian of the negative log-likelihood

loglik

log-likelihood calculated at the solution

message

message indicating if the model converged